Creates a table with sample information

Creates a table with sample information

shapeSample(X, d)

Arguments

X

A Markov Chain sample.

d

The order of the Markov chain.

Value

A table with every sequence in the sample and its frequency.

Details

The function will make a tibble with \(d+2\) columns. For each row, the first \(d+1\) columns will have a sequency of size \(d+1\) that appeared in the sample. The last column, called Nxa, will contain the number of times each of these sequences appeared in the sample.

The algorithm to obtain a tibble with all sequences that appear in the sample consists in generating d1 smaller matrices and grouping then by rows. For exemple, let d=3 and the sample X be a sequence of size n=13, where \(X=(x_1,x_2,x_3,x_4,x_5,x_6,x_7,x_8,x_9,x_{10},x_{11},x_{12},x_{13})\).

For the first matrix beggin from de element \(x_1\).

\(\begin{bmatrix}x_1 & x_2 & x_3 & x_4\\ x_5 & x_6 & x_7 & x_8\\ x_9 & x_{10} & x_{11} & x_{12}\end{bmatrix}\)

For the second matrix discard the element \(x_1\) and beggin from element \(x_2\). Than attach it to the last matrix.

\(\begin{bmatrix} x_1 & x_2 & x_3 & x_4\\ x_5 & x_6 & x_7 & x_8\\ x_9 & x_{10} & x_{11} & x_{12}\end{bmatrix}\\\begin{bmatrix}x_2 & x_3 & x_4 & x_5\\ x_6 & x_7 & x_8 & x_9\\ x_{10} & x_{11} & x_{12} & x_{13}\end{bmatrix}\)

Repeat d1 times, for the last matrix discard \(x_{d}\) and beggin from \(x_{d1}\).

\(\begin{bmatrix}x_1 & x_2 & x_3 & x_4\\ x_5 & x_6 & x_7 & x_8\\ x_9 & x_{10} & x_{11} & x_{12}\\ x_2 & x_3 & x_4 & x_5\\ x_6 & x_7 & x_8 & x_9\\ x_{10} & x_{11} & x_{12} & x_{13}\\ \vdots & & & \\ x_4 & x_5 & x_6 & x_7\\ x_8 & x_9 & x_{10} & x_{11} \end{bmatrix}\)

Note that, for every loop, the last elements of the sample are discarted, if needed, to fit the matrices.

Examples

shapeSample(c(1,2,2,1,2,1,1,2,1,2),3)
#> # A tibble: 7 x 5
#>      x3    x2    x1     a   Nxa
#>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     1     1     2     1     1
#> 2     1     2     1     1     1
#> 3     1     2     1     2     1
#> 4     1     2     2     1     1
#> 5     2     1     1     2     1
#> 6     2     1     2     1     1
#> 7     2     2     1     2     1
shapeSample(c(0,2,0,2,0,2,1,1,0,0,1,2,1,2,1),4)
#> # A tibble: 11 x 6
#>       x4    x3    x2    x1     a   Nxa
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     0     0     1     2     1     1
#>  2     0     1     2     1     2     1
#>  3     0     2     0     2     0     1
#>  4     0     2     0     2     1     1
#>  5     0     2     1     1     0     1
#>  6     1     0     0     1     2     1
#>  7     1     1     0     0     1     1
#>  8     1     2     1     2     1     1
#>  9     2     0     2     0     2     1
#> 10     2     0     2     1     1     1
#> 11     2     1     1     0     0     1